lf gate: polish v1 API + python client for ship
Fix datetime serialization in CLI JSON output (model_dump mode="json"),
remove dead stimulus param from run_wave (server doesn't accept it),
add missing type hint on Client.wave_logs, fix lfq create examples in
docs to use positional args, remove dead Rust code (api_error_with_param),
cargo fmt, handle .paused case in Concerto views.
