+++ python/loopflow/client.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import os
+from collections.abc import Iterator
 from typing import Any, Optional
 
 import httpx
@@ -102,7 +103,6 @@ class Client:
         flow: Optional[str] = None,
         direction: Optional[list[str]] = None,
         area: Optional[list[str]] = None,
-        stimulus: Optional[Stimulus] = None,
     ) -> dict[str, Any]:
         body: dict[str, Any] = {}
         if flow is not None:
@@ -111,8 +111,6 @@ class Client:
             body["direction"] = direction
         if area is not None:
             body["area"] = area
-        if stimulus is not None:
-            body["stimulus"] = stimulus.model_dump(exclude_none=True)
         return self._request_json("POST", f"/v1/waves/{name_or_id}/run", json=body)
 
     def stop_wave(self, name_or_id: str) -> dict[str, Any]:
@@ -157,7 +155,7 @@ class Client:
         data = payload.get("data", [])
         return [WaveRun.model_validate(item) for item in data]
 
-    def wave_logs(self, name_or_id: str):
+    def wave_logs(self, name_or_id: str) -> Iterator[str]:
         try:
             with self._client.stream(
                 "GET",
